[electron] Test codecs from ffmpeg
Electron's ffmpeg can bundle proprietary codecs, which we want to avoid
having.

This commit adds a native module that will be built and run on the
client where the `@theia/electron` package will be installed.

The program will link at runtime against Electron's ffmpeg shared
library and scan the available codecs, failing if a proprietary one is
found.

Note: On windows, node-gyp's path resolution is broken, so I had to
create a js file to correctly resolve node-gyp's cli.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>